requirements-dev.txt
-r requirements.txt

ruff
mypy
pytest
pytest-cov
PyYAML
types-PyYAML